+++ engine/renderer/tr_animation.cpp
@@ -45,7 +45,7 @@ static int R_MDRCullModel( mdrHeader_t *header, const trRefEntity_t *ent ) {
 	mdrFrame_t	*oldFrame, *newFrame;
 	int			i, frameSize;
 
-	frameSize = (size_t)( &((mdrFrame_t *)0)->bones[ header->numBones ] );
+	frameSize = (int)( (size_t)( &((mdrFrame_t *)0)->bones[ header->numBones ] ) );
 	
 	// compute frame pointers
 	newFrame = ( mdrFrame_t * ) ( ( byte * ) header + header->ofsFrames + frameSize * ent->e.frame);
@@ -143,7 +143,7 @@ static int R_MDRComputeFogNum( mdrHeader_t *header, const trRefEntity_t *ent ) {
 		return 0;
 	}
 	
-	frameSize = (size_t)( &((mdrFrame_t *)0)->bones[ header->numBones ] );
+	frameSize = (int)( (size_t)( &((mdrFrame_t *)0)->bones[ header->numBones ] ) );
 
 	// FIXME: non-normalized axis issues
 	mdrFrame = ( mdrFrame_t * ) ( ( byte * ) header + header->ofsFrames + frameSize * ent->e.frame);
@@ -343,7 +343,7 @@ void RB_MDRSurfaceAnim( mdrSurface_t *surface )
 
 	header = (mdrHeader_t *)((byte *)surface + surface->ofsHeader);
 
-	frameSize = (size_t)( &((mdrFrame_t *)0)->bones[ header->numBones ] );
+	frameSize = (int)( (size_t)( &((mdrFrame_t *)0)->bones[ header->numBones ] ) );
 
 	frame = (mdrFrame_t *)((byte *)header + header->ofsFrames +
 		backEnd.currentEntity->e.frame * frameSize );
